Issue #202: Reduced the file length of LocalVariableAssignedMultipleTimes
